requirements.txt
##graphsampling
numpy>=1.26.0,<3
opencv-python>=4.8.0
svgpathtools>=1.6.1
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.66.0

##cas
fastapi
uvicorn
pydantic>=2
httpx
python-dotenv
sympy
openai>=1.40.0


